conspiracy: parse: Add interface for parsing ports
SRFI-41 streams are convenient for returning "lazy" data, like syntax
objects parsed from a stream. Using streams enforces the following:

- Elements of the stream are only computed once, so (e.g.) syntax
  objects are always returned in the same order in the stream.
- Syntax errors in the port will not raise an error until they are
  actually requested.
- Work reading/parsing the input is delayed until it is actually
  required.

Add a function for converting an arbitrary port to a stream of syntax
objects. Also add a wrapper that converts a string into a stream of
syntax objects, which is useful for testing.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>