conspiracy: parse: Add interface for parsing files
The (future) linter will most commonly (exclusively?) lint files. Add
a wrapper around PORT->SYNTAX-STREAM that parses a file into a syntax
stream.

This also adds a keyword argument CLOSE? to PORT->SYNTAX-STREAM that
indicates the port should be closed at the end of the stream. Using a
stream means that we can't use a context manager to open the file
(since the stream needs the open port until it is finished), so the
stream itself needs to handle closing the port when it is a file.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>