conspiracy: lint: Initial revision
Implement the core of the linter, which matches "lints" against syntax
objects, returning context information about the failures. Lint
failures are returned as a stream while reading/parsing from some
input (string, file, or other port). These failures can be printed
using the DISPLAY-LINT-FAILURE function, which also prints the line of
code that matched the lint.

Only a handful of simple lints are included; most of these are for
reminding me of better patterns in code.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>