conspiracy: unification: Add tail variables for binding
Tail variables are constructed using VAR* and match against the entire
tail of a list, e.g. `(foo ,(var* x)) unified with '(foo bar baz) will
result in the substitution '(bar baz) for x.

This is useful for binding more complex structures where the number of
elements are not known in advanced.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>