conspiracy: lint: Add filter expressions
After matching a pattern, a lint may filter out the match with a
lambda that takes the substitution as an argument. This is primarily
useful for things where unification is not powerful enough (for
example, the filter may check the length of variable names); in the
short-term, I'll likely also use this for things my unification cannot
yet handle but may in the future (e.g. inverse matches).

By default, the filter will always return #T, meaning all pattern
matches will always produce a lint.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>